src/server/requirements.txt
flask
bootstrap-flask
ffmpeg-python